RSPY-488: Splitted staging tests into smaller files
