+++ services/staging/tests/conftest.py
@@ -65,6 +65,50 @@ S3_EXPIRATION_BUCKET_CSV_FILE = osp.join(RESOURCES_FOLDER, "expiration_bucket.cs
 TEST_DETAIL = "Test detail"
 
 
+EXPECTED_JOBS_TEST = [
+    {
+        "jobID": "job_1",
+        "status": "running",
+        "type": "process",
+        "progress": 0.0,
+        "message": "Test detail",
+        "created": datetime(2024, 1, 1, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "updated": datetime(2024, 1, 1, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "processID": "staging",
+    },
+    {
+        "jobID": "job_2",
+        "status": "running",
+        "type": "process",
+        "progress": 55.0,
+        "message": "Test detail",
+        "created": datetime(2024, 1, 2, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "updated": datetime(2024, 1, 2, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "processID": "staging",
+    },
+    {
+        "jobID": "job_3",
+        "status": "running",
+        "type": "process",
+        "progress": 15.0,
+        "message": "Test detail",
+        "created": datetime(2024, 1, 3, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "updated": datetime(2024, 1, 3, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "processID": "staging",
+    },
+    {
+        "jobID": "job_4",
+        "status": "successful",
+        "type": "process",
+        "progress": 100.0,
+        "message": "Test detail",
+        "created": datetime(2024, 1, 4, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "updated": datetime(2024, 1, 4, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
+        "processID": "staging",
+    },
+]
+
+
 @pytest.fixture(name="set_db_env_var")
 def set_db_env_var_fixture(monkeypatch):
     """Fixture to set environment variables for simulating the mounting of
