+++ services/staging/tests/test_staging_endpoints.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Test staging module."""
+"""Test staging module endpoints."""
+
 import copy
-import os
 from datetime import datetime
 
 import pytest
@@ -24,11 +24,8 @@ from rs_server_staging.main import (
     app_lifespan,
     format_job_data,
     format_jobs_data,
-    get_config_contents,
-    init_db,
     init_pygeoapi,
 )
-from sqlalchemy.exc import SQLAlchemyError
 from starlette.status import (
     HTTP_200_OK,
     HTTP_201_CREATED,
@@ -37,183 +34,7 @@ from starlette.status import (
     HTTP_500_INTERNAL_SERVER_ERROR,
 )
 
-expected_jobs_test = [
-    {
-        "jobID": "job_1",
-        "status": "running",
-        "type": "process",
-        "progress": 0.0,
-        "message": "Test detail",
-        "created": datetime(2024, 1, 1, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "updated": datetime(2024, 1, 1, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "processID": "staging",
-    },
-    {
-        "jobID": "job_2",
-        "status": "running",
-        "type": "process",
-        "progress": 55.0,
-        "message": "Test detail",
-        "created": datetime(2024, 1, 2, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "updated": datetime(2024, 1, 2, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "processID": "staging",
-    },
-    {
-        "jobID": "job_3",
-        "status": "running",
-        "type": "process",
-        "progress": 15.0,
-        "message": "Test detail",
-        "created": datetime(2024, 1, 3, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "updated": datetime(2024, 1, 3, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "processID": "staging",
-    },
-    {
-        "jobID": "job_4",
-        "status": "successful",
-        "type": "process",
-        "progress": 100.0,
-        "message": "Test detail",
-        "created": datetime(2024, 1, 4, 12, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "updated": datetime(2024, 1, 4, 13, 0, 0).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        "processID": "staging",
-    },
-]
-
-
-class TestInitDb:
-    """Class to group tests for the init_db function"""
-
-    def test_init_db_success(self, set_db_env_var, mocker):  # pylint: disable=unused-argument
-        """Test that the database initialization completes successfully."""
-
-        # Mock pygeoapi functions
-        mock_engine = mocker.Mock()
-        mocker.patch("pygeoapi.process.manager.postgresql.get_table_model", return_value=mocker.Mock())
-        mock_get_engine = mocker.patch("pygeoapi.process.manager.postgresql.get_engine", return_value=mock_engine)
-        mock_get_engine = mocker.patch("rs_server_staging.main.get_engine", return_value=mock_engine)
-        mock_metadata = mocker.patch("rs_server_staging.main.Base.metadata.create_all")
-        mocker.patch("rs_server_staging.main.api", init_pygeoapi())
-
-        # Act: Call the function
-        init_db()
-
-        # Assert: Check that create_engine and create_all were called correctly
-        mock_get_engine.assert_called_once_with(  # nosec hardcoded_password_funcarg
-            host="localhost",
-            port=5500,
-            database="rspy_pytest",
-            user="postgres",
-            password="password",
-        )
-        mock_metadata.assert_called_once_with(bind=mock_engine)
-
-    def test_init_db_missing_env_variable(self, mocker):
-        """Test that the function raises an error when environment variables are missing."""
-        # Mock environment variables to be incomplete
-        mocker.patch.dict("os.environ", {}, clear=True)
-
-        # Act & Assert: Check that an exception is raised for missing port environment variable
-        with pytest.raises(KeyError, match="POSTGRES_HOST"):
-            init_pygeoapi()
-
-    def test_init_db_sqlalchemy_error(self, set_db_env_var, mocker):  # pylint: disable=unused-argument
-        """Test that the function raises an error when SQLAlchemy fails."""
-
-        # Mock SQLAlchemy create_engine to raise an error
-        mocker.patch("rs_server_staging.main.api", init_pygeoapi())
-        mocker.patch("pygeoapi.process.manager.postgresql.get_engine", side_effect=SQLAlchemyError("Database error"))
-
-        # Act & Assert: Check that a RuntimeError is raised
-        with pytest.raises(SQLAlchemyError):
-            init_db(timeout=0)
-
-    def test_get_config_contents_success(self, set_db_env_var):  # pylint: disable=unused-argument
-        """Test that the manager definition is correctly retrieved and placeholders are replaced."""
-
-        # Act: Call the function
-        result = get_config_contents()
-
-        # Assert: Validate the updated connection dictionary
-        assert result["manager"]["connection"] == {
-            "host": os.environ["POSTGRES_HOST"],
-            "port": int(os.environ["POSTGRES_PORT"]),
-            "database": os.environ["POSTGRES_DB"],
-            "user": os.environ["POSTGRES_USER"],
-            "password": os.environ["POSTGRES_PASSWORD"],
-        }
-
-    def test_get_config_contents_invalid_definition(self, mocker):
-        """Test that the function raises an error when the manager definition is invalid."""
-        # Mock the api.config.get method to return an invalid configuration
-        mock_api_config = mocker.patch("rs_server_staging.main.api.config", autospec=True)
-        mock_api_config.get.return_value = {"connection": None}
-
-        # Act & Assert: Check that a RuntimeError is raised
-        with pytest.raises(RuntimeError, match="Error reading the manager definition for pygeoapi PostgreSQL Manager"):
-            init_db()
-
-
-def test_format_job_data(mock_jobs):
-    """
-    Check the behavior of the method that format the output of the job information returned by
-    the PostgresSQL database
-    """
-    # ----- Check that the right exception is raised if the input job data
-    # doesn't have the right format
-    wrong_mock_job = copy.deepcopy(mock_jobs[0])
-    wrong_mock_job["wrong_attribute"] = wrong_mock_job.pop("identifier")
-    with pytest.raises(Exception) as excinfo:
-        format_job_data(wrong_mock_job)
-    assert "attribute 'identifier' is missing" in str(excinfo.value)
-
-    # ----- Check that the input job is well formatted
-    mock_job = copy.deepcopy(mock_jobs[0])
-    expected_response = expected_jobs_test[0]
-    assert format_job_data(mock_job) == expected_response
-
-
-def test_format_jobs_data(mock_jobs):
-    """
-    Check the behavior of the method that format the output of the job information returned by
-    the PostgresSQL database
-    """
-    expected_response = {
-        "jobs": expected_jobs_test,
-        "links": [
-            {
-                "href": "string",
-                "rel": "service",
-                "type": "application/json",
-                "hreflang": "en",
-                "title": "List of jobs",
-            },
-        ],
-    }
-    # ----- Check that the right exception is raised if the input jobs is something else than a dictionary
-    with pytest.raises(Exception) as excinfo:
-        format_jobs_data("wrong_data")  # type: ignore
-    assert "Expected a dictionary as input" in str(excinfo.value)
-
-    # ----- Check that the right exception is raised if the input job doesn't have the required 'jobs' attributes
-    with pytest.raises(Exception) as excinfo:
-        format_jobs_data({"attr1": "val1", "attr2": "val2"})
-    assert "Invalid format for input jobs: missing 'jobs' key" in str(excinfo.value)
-
-    # ----- Check that the input job is well formatted if the input has the correct format
-    mock_jobs = {
-        "jobs": copy.deepcopy(mock_jobs),
-        "links": [
-            {
-                "href": "string",
-                "rel": "service",
-                "type": "application/json",
-                "hreflang": "en",
-                "title": "List of jobs",
-            },
-        ],
-    }
-    assert format_jobs_data(mock_jobs) == expected_response
+from .conftest import EXPECTED_JOBS_TEST
 
 
 @pytest.mark.asyncio
@@ -335,8 +156,8 @@ async def test_get_jobs_endpoint(
             HTTP_404_NOT_FOUND,
             "Job with ID non_existing_id not found",
         ),
-        (expected_jobs_test[0], HTTP_500_INTERNAL_SERVER_ERROR, "'type' is a required property"),
-        *[(job, HTTP_200_OK, job) for job in expected_jobs_test],
+        (EXPECTED_JOBS_TEST[0], HTTP_500_INTERNAL_SERVER_ERROR, "'type' is a required property"),
+        *[(job, HTTP_200_OK, job) for job in EXPECTED_JOBS_TEST],
     ],
 )
 async def test_get_job(
@@ -419,11 +240,11 @@ async def test_get_job(
             "Job with ID non_existing_id not found",
         ),
         (
-            expected_jobs_test[0],
+            EXPECTED_JOBS_TEST[0],
             HTTP_500_INTERNAL_SERVER_ERROR,
             "{'status': 'format'} is not valid under any of the given schemas",
         ),
-        *[(job, HTTP_200_OK, job["status"]) for job in expected_jobs_test],
+        *[(job, HTTP_200_OK, job["status"]) for job in EXPECTED_JOBS_TEST],
     ],
 )
 async def test_get_job_result(
@@ -500,8 +321,8 @@ async def test_get_job_result(
     "expected_job, expected_status, expected_response",
     [
         ({"jobID": "non_existing_id"}, HTTP_404_NOT_FOUND, "Job with ID non_existing_id not found"),
-        (expected_jobs_test[0], HTTP_500_INTERNAL_SERVER_ERROR, "'type' is a required property"),
-        *[(job, HTTP_200_OK, f"Job {job['jobID']} deleted successfully") for job in expected_jobs_test],
+        (EXPECTED_JOBS_TEST[0], HTTP_500_INTERNAL_SERVER_ERROR, "'type' is a required property"),
+        *[(job, HTTP_200_OK, f"Job {job['jobID']} deleted successfully") for job in EXPECTED_JOBS_TEST],
     ],
 )
 async def test_delete_job_endpoint(
@@ -736,14 +557,14 @@ async def test_execute_staging(
 
     # ----- Test case where both staging body and response are compliant with ogc
     mock_db_table.get_job.return_value = next(
-        job for job in mock_jobs if job["identifier"] == expected_jobs_test[0]["jobID"]
+        job for job in mock_jobs if job["identifier"] == EXPECTED_JOBS_TEST[0]["jobID"]
     )
     # Patch app.extra with the mock db_table
     staging_client.app.extra["process_manager"] = mock_db_table
     staging_client.app.extra["dask_cluster"] = None
     response = staging_client.post(f"/processes/{resource_name}/execution", json=valid_staging_body)
     assert response.status_code == HTTP_201_CREATED
-    assert response.json() == expected_jobs_test[0]
+    assert response.json() == EXPECTED_JOBS_TEST[0]
 
     # ----- Test case where we have a staging response which is uncompliant with ogc
     wrong_ogc_mock_jobs = copy.deepcopy(mock_jobs)
@@ -752,7 +573,7 @@ async def test_execute_staging(
         job.pop("type")
     # Remove required ogc attribute "type"
     mock_db_table.get_job.return_value = next(
-        job for job in wrong_ogc_mock_jobs if job["identifier"] == expected_jobs_test[0]["jobID"]
+        job for job in wrong_ogc_mock_jobs if job["identifier"] == EXPECTED_JOBS_TEST[0]["jobID"]
     )
     staging_client.app.extra["process_manager"] = mock_db_table
     response = staging_client.post(f"/processes/{resource_name}/execution", json=valid_staging_body)
